some more friend fixes
