+++ config/authConfig.js
@@ -1,5 +1,6 @@
 module.exports = {
     facebook_app_id:"2244373479132911",
     facebook_app_secret: "38be54e5144a107f348803624bec95b2",
-    callback_url : "https://recca.herokuapp.com/login/facebook/callback" 
+    callback_url : "https://recca.herokuapp.com/login/facebook/callback",
+    sessionSecret : 'Recca Raccoon' 
 }
