+++ routes/api/index.js
@@ -5,6 +5,7 @@ const spotifyRoutes = require("./spotify");
 const vgRoutes = require("./vg");
 const omdbRoutes = require("./omdb");
 const postRoutes = require("./post");
+const friendRoutes = require("./friend")
 
 router.use("/media", mediaRoutes);
 router.use("/user", userRoutes);
@@ -12,6 +13,6 @@ router.use("/spotify", spotifyRoutes);
 router.use("/vg", vgRoutes);
 router.use("/omdb", omdbRoutes);
 router.use("/post", postRoutes);
-
+router.use("/friend", friendRoutes);
 
 module.exports = router;
