+++ routes/api/user.js
@@ -8,18 +8,6 @@ router.route("/create")
 //route for finding user by id
 router.route("/find")
 .get(userController.findUser)
-.put(userController.removeFriend);
-
-//route for finding user by username
-router.route("/find/user")
-.post(userController.userByName);
-
-//route for friend requests
-router.route("/friend")
-.get(userController.pendingRequest)
-.post(userController.newFriendRequest)
-.put(userController.handleFriendRequest)
-.delete(userController.removeFriend);
 
 router.route("/feed/")
 .get(userController.getFeed);
