+++ routes/index.js
@@ -14,4 +14,5 @@ router.route("/logout").get(function(req, res){
   res.send("You're logged out")
 });
 
+
 module.exports = router;
