+++ server.js
@@ -8,12 +8,13 @@ const mongoose = require("mongoose")
 const connection = mongoose.connect(process.env.MONGODB_URI || "mongodb://localhost:27017/recca");
 const MongoStore = require("connect-mongo")(session);
 const passport = require("passport");
+const config = require("./config/authConfig")
 
 // Express middleware
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 app.use(session({
-  secret: 'Recco Gecco',
+  secret: config.sessionSecret,
   store: new MongoStore({
     mongooseConnection: mongoose.connection
   })
